% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file.R
\name{calc_blank_ratio}
\alias{calc_blank_ratio}
\title{Calculate the sample to average blank ratio}
\usage{
calc_blank_ratio(
  data = NULL,
  blanks = NULL,
  samples = NULL,
  ratio = 5,
  threshold = 0.8
)
}
\arguments{
\item{data}{tibble in tidy format.}

\item{blanks}{character() vector with the blank names.}

\item{samples}{character() vector with the sample names.}

\item{ratio}{numeric(1) cutoff ratio of sample/ blank.}

\item{threshold}{numeric(1) threshold.}
}
\value{
Returns a tibble in tidy with sample / blank ratio for each lipid
    species.
}
\description{
Calculate for each lipid the ratio between the sample and the
    average of all blanks.
}
\author{
Rico Derks
}
