% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{calc_cor}
\alias{calc_cor}
\title{Calculate correlation between all samples and pooled samples}
\usage{
calc_cor(data = NULL, idx_samples = NULL, idx_pools = NULL)
}
\arguments{
\item{data}{data.frame with all the data.}

\item{idx_samples}{character(), with all the sample names.}

\item{idx_pools}{character(), with all the pooled sample names.}
}
\value{
data.frame in long format for `qc_cor_plot()`.
}
\description{
Calculate correlation between all samples and pooled samples.
}
\author{
Rico Derks
}
