% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file.R
\name{calc_rsd}
\alias{calc_rsd}
\title{Calculate RSD values of all lipids}
\usage{
calc_rsd(data = NULL, pools = NULL, cut_off = 0.3)
}
\arguments{
\item{data}{tibble in tidy format.}

\item{pools}{character() vector with the qcpool names.}

\item{cut_off}{numeric(1), the RSD cut off value.}
}
\value{
Character vector with id's.
}
\description{
Calculate RSD values of all lipids within all QCpool samples..
}
\author{
Rico Derks
}
