% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file.R
\name{clean_up}
\alias{clean_up}
\title{Clean up tibble from MS-DIAL}
\usage{
clean_up(raw_data = NULL, blanks = NULL, pools = NULL, samples = NULL)
}
\arguments{
\item{raw_data}{The tibble.}

\item{blanks}{character() vector with the name of the blanks to keep.}

\item{pools}{character() vector with the name of the pooled samples to keep.}

\item{samples}{character() vector with the name of the samples to keep.}
}
\value{
Returns a tibble
}
\description{
Clean up the columns and column names of the tibble after
    reading the MS-DIAL result files.
}
\author{
Rico Derks
}
