% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file.R
\name{filter_id}
\alias{filter_id}
\title{Calculate the sample to average blank ratio}
\usage{
filter_id(data = NULL, dot_cutoff = 50, revdot_cutoff = 50)
}
\arguments{
\item{data}{tibble in tidy format.}

\item{dot_cutoff}{numeric(1) cutoff value dot product.}

\item{revdot_cutoff}{numeric(1) reverse dot product cutoff value.}
}
\value{
Character vector with id's.
}
\description{
Calculate for each lipid the ratio between the sample and the
    average of all blanks.
}
\author{
Rico Derks
}
