% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file.R
\name{make_tidy}
\alias{make_tidy}
\title{Make the tibble in tidy format}
\usage{
make_tidy(
  data = NULL,
  blanks = NULL,
  pools = NULL,
  samples = NULL,
  omics = c("lip\\", \\"met")
)
}
\arguments{
\item{data}{The tibble.}

\item{blanks}{character() vector with the name of the blanks to keep.}

\item{pools}{character() vector with the name of the pooled samples to keep.}

\item{samples}{character() vector with the name of the samples to keep.}

\item{omics}{character(1), lipidomics or metabolomics data}
}
\value{
Returns a tibble in tidy (long) format
}
\description{
Filter the tibble to keep only the identified lipids.
}
\details{
After making the tibble in long format also some additional columns ared added.
}
\author{
Rico Derks
}
