% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{qc_cor_plot}
\alias{qc_cor_plot}
\title{Correlation heatmap}
\usage{
qc_cor_plot(data = NULL)
}
\arguments{
\item{data}{data.frame with all the data.}
}
\value{
ggplot2 object, histogram of the RSD values
}
\description{
Correlation heatmap of all samples and pooled samples.
}
\author{
Rico Derks
}
