% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file.R
\name{read_msdial}
\alias{read_msdial}
\title{Read MS-DIAl files}
\usage{
read_msdial(filename = NULL)
}
\arguments{
\item{filename}{The filename of the MS-DIAL file.}
}
\value{
Returns a tibble
}
\description{
Read the result files from MS-DIAL.
}
\details{
file
}
\author{
Rico Derks
}
