% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file.R
\name{select_identified}
\alias{select_identified}
\title{Keep only the identified lipids}
\usage{
select_identified(data = NULL, omics = c("lip", "met"))
}
\arguments{
\item{data}{The tibble.}

\item{omics}{character(1), lipidomics or metabolomics data}
}
\value{
Returns a tibble
}
\description{
Filter the tibble to keep only the identified lipids.
}
\author{
Rico Derks
}
