% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{show_class_violin}
\alias{show_class_violin}
\title{Create RSD violin plot}
\usage{
show_class_violin(data = NULL, rsd_cutoff = 0.3)
}
\arguments{
\item{data}{data.frame with the RSD data.}

\item{rsd_cutoff}{numeric(1), the RSD cut off value.}
}
\value{
ggplot2 object, violin plot of the RSD values per class.
}
\description{
Create RSD violin plot per class.
}
\details{
data should contain the columns RSD, class and polarity.
}
\author{
Rico Derks
}
