% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{show_trend_plot}
\alias{show_trend_plot}
\title{Create trend plot}
\usage{
show_trend_plot(data = NULL)
}
\arguments{
\item{data}{data.frame from `calc_trend()`.}
}
\value{
ggplot2 object
}
\description{
Create trend plot of the pooled samples to visualize the trend during the
measurements..
}
\author{
Rico Derks
}
