% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{trend_plot}
\alias{trend_plot}
\title{Create trend plot}
\usage{
trend_plot(trend_data = NULL, type = c("batch", "overall"))
}
\arguments{
\item{trend_data}{data.frame containing all trend related data.}

\item{type}{character(1) do you want the overall trend plot or per batch.}
}
\value{
trend plot as ggplot2 object.
}
\description{
Create a trend plot based on the pooled samples in the data.
}
\author{
Rico Derks
}
