% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{calc_trend}
\alias{calc_trend}
\title{Calculate (log2) fold change of the pooled samples}
\usage{
calc_trend(data = NULL, idx_pools = NULL)
}
\arguments{
\item{data}{data.frame with all the data.}

\item{idx_pools}{character(), with all the pooled sample names.}
}
\value{
data.frame in long format for `qc_trend_plot()`.
}
\description{
Calculate (log2) fold change of the pooled samples. The reference is the first
pooled sample.
}
\author{
Rico Derks
}
