% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc.R
\name{calc_trend}
\alias{calc_trend}
\title{Get trend data}
\usage{
calc_trend(pool_data = NULL, order_column = NULL)
}
\arguments{
\item{pool_data}{data.frame with the pooled data.}

\item{order_column}{character(1) column name of the acquisition order column.}
}
\value{
data.frame with the data for the trend plot
}
\description{
Get the trend data for the trend plot.
}
\author{
Rico Derks
}
